Rename FromBytes to FromHeaderValue
To show its intended purpose is to parse header values. Also move it to
the header module.
